Reviewed code changes, added code review comments and changes delete method to include NULL check

+++ EmployeeReimbursmentSystem/src/main/java/com/revature/repository/ReimbursementStatusDAOImpl.java
@@ -44,11 +44,13 @@ public class ReimbursementStatusDAOImpl implements ReimbursementStatusDAO{
 
 	@Override
 	public List<ReimbursementStatus> selectAll() {
+		//TODO: please use plural in log entries and variables names
 		logger.info("Getting all reimbursement status from database....");
 		
 		Session session  = HibernateUtil.getSession();
 		
 		List<ReimbursementStatus> reimbursementStatus = session.createQuery("from Reimbursement", ReimbursementStatus.class).list();
+		//TODO: size method is missing
 		logger.info("Reimbursement status list retireved! Size: " + reimbursementStatus);
 		return reimbursementStatus;
 	}
@@ -71,22 +73,30 @@ public class ReimbursementStatusDAOImpl implements ReimbursementStatusDAO{
 	}
 
 	@Override
-	public boolean delete(ReimbursementStatus reimbursementStatus) {
-		logger.info("Deleting reimbursement status. Reimbursement status info: " + reimbursementStatus);
+	public boolean delete(int id) {
+		ReimbursementStatus reimbursementStatus = sellectById(id);
 		
-		Session session = HibernateUtil.getSession();
+		if (reimbursementStatus != null) {
+			logger.info("Deleting reimbursement status. Reimbursement status info: " + reimbursementStatus);
+			
+			Session session = HibernateUtil.getSession();
+			
+			Transaction tx = session.beginTransaction();
+			
+			session.clear();
+			
+			session.delete(reimbursementStatus);
+			
+			tx.commit();
+			
+			logger.info("Successfully deleted the reimbursement status!");
+			
+			return true;
+		} else {
+			logger.info("Delete failed: Unable to find a reimbursement status with id: " + id);
+			return false;	
+		}
 		
-		Transaction tx = session.beginTransaction();
-		
-		session.clear();
-		
-		session.delete(reimbursementStatus);
-		
-		tx.commit();
-		
-		logger.info("Successfully deleted the reimbursement status!");
-		
-		return true;
 	}
 
 }
